+++ src/clients/bridges/index.ts
@@ -1,6 +1,6 @@
 export * from "./BaseAdapter";
 export * from "./AdapterManager";
-export * from "./OptimismAdapter";
+export * from "./op-stack/optimism/OptimismAdapter";
 export * from "./ArbitrumAdapter";
 export * from "./PolygonAdapter";
 export * from "./CrossChainTransferClient";
