+++ src/clients/bridges/op-stack/OpStackAdapter.ts
@@ -1,169 +1,17 @@
 import assert from "assert";
-import {
-  Contract,
-  BigNumber,
-  ZERO_ADDRESS,
-  paginatedEventQuery,
-  BigNumberish,
-  TransactionResponse,
-  Event,
-  checkAddress,
-  Signer,
-  EventSearchConfig,
-  Provider,
-  ethers,
-} from "../../utils";
-import { spreadEventWithBlockNumber, assign, winston } from "../../utils";
-import { SpokePoolClient } from "../../clients";
-import { BaseAdapter } from "./";
-import { SortableEvent } from "../../interfaces";
-import { OutstandingTransfers } from "../../interfaces";
-import { CONTRACT_ADDRESSES } from "../../common";
+import { Contract, BigNumber, BigNumberish, TransactionResponse, Event, checkAddress, ethers } from "../../../utils";
+import { spreadEventWithBlockNumber, assign, winston } from "../../../utils";
+import { SpokePoolClient } from "../..";
+import { BaseAdapter } from "..";
+import { SortableEvent } from "../../../interfaces";
+import { OutstandingTransfers } from "../../../interfaces";
+import { CONTRACT_ADDRESSES } from "../../../common";
 import { constants } from "@across-protocol/sdk-v2";
+import { OpStackBridge } from "./OpStackBridgeInterface";
+import { WethBridge } from "./WethBridge";
+import { DefaultERC20Bridge } from "./DefaultErc20Bridge";
+
 const { TOKEN_SYMBOLS_MAP } = constants;
-
-export interface TransactionDetails {
-  readonly contract: Contract;
-  readonly method: string;
-  readonly args: any[];
-}
-
-export interface EventSearchDetails {
-  readonly address: string;
-  readonly eventName: string;
-  readonly queryParams: any[];
-}
-
-export interface OpStackBridge {
-  readonly l1Gateway: string;
-  constructL1ToL2Txn(
-    toAddress: string,
-    l1Token: string,
-    l2Token: string,
-    amount: BigNumber,
-    l2Gas: number
-  ): TransactionDetails;
-  queryL1BridgeInitiationEvents(l1Token: string, fromAddress: string, eventConfig: EventSearchConfig): Promise<Event[]>;
-  queryL2BridgeFinalizationEvents(
-    l1Token: string,
-    fromAddress: string,
-    eventConfig: EventSearchConfig
-  ): Promise<Event[]>;
-}
-
-class DefaultERC20Bridge implements OpStackBridge {
-  private readonly l1Bridge: Contract;
-  private readonly l2Bridge: Contract;
-
-  constructor(private l2chainId: number, hubChainId: number, l1Signer: Signer, l2SignerOrProvider: Signer | Provider) {
-    const { address: l1Address, abi: l1Abi } = CONTRACT_ADDRESSES[hubChainId][`ovmStandardBridge_${l2chainId}`];
-    this.l1Bridge = new Contract(l1Address, l1Abi, l1Signer);
-
-    const { address: l2Address, abi: l2Abi } = CONTRACT_ADDRESSES[l2chainId].ovmStandardBridge;
-    this.l2Bridge = new Contract(l2Address, l2Abi, l2SignerOrProvider);
-  }
-
-  get l1Gateway() {
-    return this.l1Bridge.address;
-  }
-
-  constructL1ToL2Txn(
-    toAddress: string,
-    l1Token: string,
-    l2Token: string,
-    amount: BigNumber,
-    l2Gas: number
-  ): TransactionDetails {
-    return {
-      contract: this.l1Bridge,
-      method: "depositERC20",
-      args: [l1Token, l2Token, amount, l2Gas, "0x"],
-    };
-  }
-
-  queryL1BridgeInitiationEvents(
-    l1Token: string,
-    fromAddress: string,
-    eventConfig: EventSearchConfig
-  ): Promise<Event[]> {
-    return paginatedEventQuery(
-      this.l1Bridge,
-      this.l1Bridge.filters.ETHDepositInitiated(undefined, fromAddress),
-      eventConfig
-    );
-  }
-
-  queryL2BridgeFinalizationEvents(
-    l1Token: string,
-    fromAddress: string,
-    eventConfig: EventSearchConfig
-  ): Promise<Event[]> {
-    return paginatedEventQuery(
-      this.l2Bridge,
-      this.l2Bridge.filters.DepositFinalized(ZERO_ADDRESS, undefined, fromAddress),
-      eventConfig
-    );
-  }
-}
-
-class WethBridge implements OpStackBridge {
-  private readonly l1Bridge: Contract;
-  private readonly l2Bridge: Contract;
-  private readonly atomicDepositor: Contract;
-
-  constructor(private l2chainId: number, hubChainId: number, l1Signer: Signer, l2SignerOrProvider: Signer | Provider) {
-    const { address: l1Address, abi: l1Abi } = CONTRACT_ADDRESSES[hubChainId][`ovmStandardBridge_${l2chainId}`];
-    this.l1Bridge = new Contract(l1Address, l1Abi, l1Signer);
-
-    const { address: l2Address, abi: l2Abi } = CONTRACT_ADDRESSES[l2chainId].ovmStandardBridge;
-    this.l2Bridge = new Contract(l2Address, l2Abi, l2SignerOrProvider);
-
-    const { address: atomicDepositorAddress, abi: atomicDepositorAbi } = CONTRACT_ADDRESSES[hubChainId].atomicDepositor;
-    this.atomicDepositor = new Contract(atomicDepositorAddress, atomicDepositorAbi, l1Signer);
-  }
-
-  get l1Gateway() {
-    return this.atomicDepositor.address;
-  }
-
-  constructL1ToL2Txn(
-    toAddress: string,
-    l1Token: string,
-    l2Token: string,
-    amount: BigNumber,
-    l2Gas: number
-  ): TransactionDetails {
-    return {
-      contract: this.atomicDepositor,
-      method: "bridgeWethToOvm",
-      args: [toAddress, amount, l2Gas, this.l2chainId],
-    };
-  }
-
-  queryL1BridgeInitiationEvents(
-    l1Token: string,
-    fromAddress: string,
-    eventConfig: EventSearchConfig
-  ): Promise<Event[]> {
-    return paginatedEventQuery(
-      this.l1Bridge,
-      this.l1Bridge.filters.ERC20DepositInitiated(l1Token, undefined, fromAddress),
-      eventConfig
-    );
-  }
-
-  queryL2BridgeFinalizationEvents(
-    l1Token: string,
-    fromAddress: string,
-    eventConfig: EventSearchConfig
-  ): Promise<Event[]> {
-    return paginatedEventQuery(
-      this.l2Bridge,
-      this.l2Bridge.filters.DepositFinalized(l1Token, undefined, fromAddress),
-      eventConfig
-    );
-  }
-}
 
 export class OpStackAdapter extends BaseAdapter {
   public l2Gas: number;
