+++ src/clients/bridges/op-stack/optimism/index.ts
@@ -1,0 +1,2 @@
+// Note: the custom bridges _do not_ need to be exported.
+export * from "./OptimismAdapter";
